refactor date parsers
