+++ scratch.js
@@ -29,7 +29,9 @@ spacetime.extend(require('./plugins/dst/src/index.js'))
 // console.log(d.format('nice-day'))
 // d = d.add(1, 'weekend')
 // console.log(d.format('nice-day'))
-let a = spacetime('3 February 2003', 'Canada/Eastern')
+let a = spacetime("8th February '86", 'Canada/Eastern')
+// let a = spacetime('1 january 2020', 'Canada/Eastern')
+// let a = spacetime('2016-feb-01', 'Canada/Eastern')
 // let d = spacetime('1992-03-21T18:00:00.000+11:30')
 console.log(a.format('nice'))
 // d = d.add(1, 'weekend')
