+++ src/input/formats/index.js
@@ -1,0 +1,6 @@
+module.exports = [].concat(
+  require('./01-ymd'),
+  require('./02-mdy'),
+  require('./03-dmy'),
+  require('./04-misc')
+)
