+++ src/input/formats/parseOffset.js
